engines/baseline: drop int16-to-float casts

+++ src/engines/baseline_engine.cpp
@@ -13,8 +13,8 @@ void TraverseRecursive(const BspData& bsp, const Frustum& frustum, int node_inde
   }
   stats.visited_nodes++;
   const auto& node = bsp.nodes[static_cast<size_t>(node_index)];
-  const Vec3 mins{static_cast<float>(node.mins[0]), static_cast<float>(node.mins[1]), static_cast<float>(node.mins[2])};
-  const Vec3 maxs{static_cast<float>(node.maxs[0]), static_cast<float>(node.maxs[1]), static_cast<float>(node.maxs[2])};
+  const Vec3 mins{node.mins[0], node.mins[1], node.mins[2]};
+  const Vec3 maxs{node.maxs[0], node.maxs[1], node.maxs[2]};
   if (!AabbIntersectsFrustum(frustum, mins, maxs)) {
     stats.culled_nodes++;
     return;
@@ -29,8 +29,8 @@ void TraverseRecursive(const BspData& bsp, const Frustum& frustum, int node_inde
       if (leaf_index >= 0 && leaf_index < static_cast<int>(bsp.leafs.size())) {
         stats.visited_leafs++;
         const auto& leaf = bsp.leafs[static_cast<size_t>(leaf_index)];
-        const Vec3 lmins{static_cast<float>(leaf.mins[0]), static_cast<float>(leaf.mins[1]), static_cast<float>(leaf.mins[2])};
-        const Vec3 lmaxs{static_cast<float>(leaf.maxs[0]), static_cast<float>(leaf.maxs[1]), static_cast<float>(leaf.maxs[2])};
+        const Vec3 lmins{leaf.mins[0], leaf.mins[1], leaf.mins[2]};
+        const Vec3 lmaxs{leaf.maxs[0], leaf.maxs[1], leaf.maxs[2]};
         if (AabbIntersectsFrustum(frustum, lmins, lmaxs)) {
           stats.accepted_leafs++;
         }
